Fixed sample marking error
I was alerted to a bug in the code that led ts2sls to treat the data as
if it had no observations. A change to sample marking has fixed the
problem, although in future versions, observations with missing lhs or
endogenous variables will need to be explicitly marked as not to use.